[STYLE]: Separación de listado de contribuciones en nuevo componente #15
